refactor: useSession에서 소켓 이벤트 코드 분리
