+++ frontend/src/components/sessions/create/SessionForm/ListSelectModal/CategoryTab/Category.tsx
@@ -14,7 +14,7 @@ const Category = ({ tabName, tabText }: Props) => {
       ${
         tab === tabName
           ? "text-bold-s text-green-500"
-          : "text-medium-m text-gray-400"
+          : "text-medium-l text-gray-400"
       }`}
       onClick={() => setTab(tabName)}
     >
