+++ frontend/src/components/sessions/create/SessionForm/ListSelectModal/QuestionList/index.tsx
@@ -32,7 +32,7 @@ const QuestionList = ({ page }: { page: number }) => {
       setQuestionList([]);
     }
   };
-  // const { tab } = useSessionFormStore();
+  //const { tab } = useSessionFormStore();
 
   const startIndex = (page - 1) * QUESTION_PER_PAGE;
   const endIndex = startIndex + QUESTION_PER_PAGE;
@@ -40,13 +40,13 @@ const QuestionList = ({ page }: { page: number }) => {
   return (
     <div className="mb-4 h-80 overflow-y-auto">
       <LoadingIndicator loadingState={questionLoading} />
-      {/*{data[tab].map((item, id) => {*/}
-      {/*  return (*/}
-      {/*    <div key={id}>*/}
-      {/*      <QuestionItem item={item} />*/}
-      {/*    </div>*/}
-      {/*  );*/}
-      {/*})}*/}
+      {/*data[tab].map((item, id) => {
+          return (
+            <div key={id}>
+              <QuestionItem item={item} />
+            </div>
+          );
+        })*/}
       {questionList.slice(startIndex, endIndex).map((item, id) => {
         return (
           <div key={id}>
