+++ frontend/src/hooks/type/session.d.ts
@@ -25,6 +25,7 @@ export interface ResponseMasterChanged {
 }
 
 export interface Participant {
+  id?: string;
   nickname: string;
   isHost: boolean;
 }
